tar.c: pull out functionality into a separate archive.c file
The zipfile code has copied a large amount of functions directly from
tar.c

This creates a maintenance issue where two copies have to be updated,
and kept synchronised whenever changes are made.

To avoid this issue, this commit pulls the functions out of tar.c, and
into a common archive.c file.  Later commits will replace the copied
code in zipfile.c with calls to these functions.
